pyproject.toml
[tool.poetry]
name = "designsafe"
version = "6.0"
description = ""
authors = ["DesignSafe-CI <designsafe-ci@tacc.utexas.edu>"]

[tool.poetry.dependencies]
python = "^3.11"
pytas = {git = "https://bitbucket.org/taccaci/pytas.git", tag = "v1.7.0"}
attrdict = { git = "https://github.com/DesignSafe-CI/AttrDict", rev = "83b779ee82d5b0e33be695d398162b8f2430ff33" }
Django = "^4.2"
daphne = "^4.0.0"
debugpy = "^1.8.0"
channels = "^4.0.0"
channels-redis = "^4.1.0"
cryptography = "^41.0.4"
django-cms = "^3.11.3 "
importlib-resources = "^5.4.0"
django-filer = "^2.2"
django-formtools = "2.2"
django-haystack = "3.2.1"
django-impersonate = "1.9.1"
django-treebeard = "^4.4"
django-sekizai = "2.0.0"
django-termsandconditions = "^2.0.12"
google-api-python-client = "^2.33.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.4.6"
boxsdk = "^2.9.0"
celery = "^5.3.4"
uWSGI = "^2.0.20"
uwsgitop = "^0.11"
python-magic = "^0.4.13"
rt = "^1.0.8"
PyJWT = "^1.4.2"
pymongo = "^3.9.0"
ipython = "^8.16.1"
ipdb = "^0.13.13"
elasticsearch = "^7.5.1"
elasticsearch-dsl = "^7.1.0"
dropbox = "10.6.0"
django-recaptcha2 = { git = "https://github.com/DesignSafe-CI/django-recaptcha2", tag = "v1.4.1--compat-django4" }
djangocms-snippet = "3.0.0"
django-bootstrap3 = "^23.4"
djangocms-text-ckeditor = "^5.1"
djangocms-cascade = { git = "https://github.com/DesignSafe-CI/djangocms-cascade", rev = "9e9f9e3088a0fcfdc1b27ad7e08b68390aa6a570" }
djangocms-style = "3.0.0"
djangocms-video = "3.0.0"
djangocms-picture = "^3.0.0"
djangocms-file = "3.0.0"
djangocms-forms-maintained = { git = "https://github.com/TACC/djangocms-forms", rev = "63ead9288c2ea65139124698bffc0ad01d182afa" }
djangocms-googlemap = "2.0.0"
oauth2client = "^4.1.2"
psycopg = "^3.1.12"
jsonpickle = "^3.0.2"
stone = "^3.2.1"
pymemcache = "^4.0.0"
django-ipware = "^1.1.6"
pytest = "^7.4.2"
pytest-django = "^4.5.2"
pytest-asyncio = "^0.14.0"
pytz = "^2023.3"
black = "^23.10.0"
pylint = "^2.12.2"
pylint-django = "2.5.5"
mock = "^4.0.3"
requests-mock = "^1.9.3"
pytest-mock = "^3.6.1"
pytest-cov = "^2.10.0"
ExifRead = "^2.3.2"
opf-fido = "1.4.1"
asgiref = "^3.7.2"
django-select2 = "6.3.1"
djangocms-admin-style = "~3.2.6"
pydantic = "^2.5.0"
networkx = "^3.2.1"
tapipy = "^1.6.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
